C01mix4
CLIC303 825793670 222 MB/s i7-13700KF

+++ ppmcodec/c09.c
@@ -409,7 +409,7 @@ int c09_codec(const char *srcfn, const char *dstfn)
 					{
 						cdf=freq;
 						freq=curr_CDF[sym+1];
-						if(freq>code)
+						if((unsigned)freq>code)
 							break;
 						++sym;
 					}
@@ -429,7 +429,7 @@ int c09_codec(const char *srcfn, const char *dstfn)
 					{
 						cdf=freq;
 						freq=curr_CDF[sym+1];
-						if(freq>code)
+						if((unsigned)freq>code)
 							break;
 						++sym;
 					}
@@ -449,7 +449,7 @@ int c09_codec(const char *srcfn, const char *dstfn)
 					{
 						cdf=freq;
 						freq=curr_CDF[sym+1];
-						if(freq>code)
+						if((unsigned)freq>code)
 							break;
 						++sym;
 					}
